# app/docs/swagger.yaml
openapi: 3.0.3
info:
  title: App Menu API
  description: Api of the App Menu project whose purpose is to generate all the endpoints for use from the frontend 
  termsOfService: http://innteligen.io/terms/
  contact:
    email: cristian.salas@innteligen.com
  version: 1.0.0
servers:
  - url: http://localhost:4000
    description: development environment
...
paths:
...
  /api/account/{accountId}:
    get:
      tags:
        - Account
      summary: Find account by ID
      description: Returns a single account
      parameters:
        - name: accountId
          in: path
          description: ID of account to return
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'          
        '400':
          description: Invalid ID supplied
        '404':
          description: Account not found
    put:
      tags:
        - Account
      summary: Update account
      description: Update user account.
      parameters:
        - name: accountId
          in: path
          description: ID of account to update
          required: true
          schema:
            type: string
      requestBody:
        description: Update an existent account
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Account'
      responses:
        default:
          description: successful operation
          
    delete:
      tags:
        - Account
      summary: Delete account
      description: Delete user account.
      operationId: deleteUser
      parameters:
        - name: accountId
          in: path
          description: ID of account to delete
          required: true
          schema:
            type: string
      responses:
        '400':
          description: Invalid account supplied
        '404':
          description: Account not found
components:
  schemas:
    Account:
      type: object
      properties:
        first_name:
          type: string
        last_name:
          type: string
        user_name:
          type: string
        password:
          type: string
          format: password
        billing_address:
          type: string
        billing_city:
          type: string
        billing_country:
          type: string
        vat_id:
          type: integer
          format: int64
          example: 1234567
        company_name: 
          type: string
        restaurants:
          type: array
          items:
            type: string
          example: ["1223233565"]